move `utils` to `index.rs`
